Switch PartnerWebClient / Server names
